Fix: изменение файловой структруры /view
