+++ src/presenter/films-presenter.js
@@ -3,7 +3,7 @@ import FilmsContainer from '../view/films-container-view';
 import FilmsList from '../view/films-list-view';
 import FilmsShowMoreBtn from '../view/films-show-more-btn-view';
 import FilmsCard from '../view/film-card-view';
-import FilmsPopup from '../view/film-popup-view';
+import FilmsPopup from '../view/film-popup';
 
 export default class FilmsPresenter {
   siteBodyElement = document.body;
