+++ src/view/film-popup/index.js
@@ -1,0 +1,25 @@
+import { createElement } from '../../render';
+import createFilmsPopupTemplate from './film-popup-view';
+
+export default class FilmsPopup {
+  constructor(film, comments) {
+    this.film = film;
+    this.comments = comments;
+  }
+
+  getTemplate() {
+    return createFilmsPopupTemplate(this.film, this.comments);
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
